requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
catboost
xgboost
Flask
dill
category_encoders
shap
uvicorn 
fastapi
pydantic
google.generativeai
python-multipart
# -e .
